move ContainerButton into identifierpage
